Squashed commit of the following:
commit c873d6af7c9cd0c03e0ddff19e6fd77a0d47717e
Author: Harmen Stoppels <harmenstoppels@gmail.com>
Date:   Wed Mar 23 10:07:53 2022 +0100

    comment

commit 861bb2ebf32323cd7ddb1fad52b1e42cc16ba938
Author: Harmen Stoppels <harmenstoppels@gmail.com>
Date:   Mon Mar 21 12:37:48 2022 +0100

    Add a test

commit 16843386badf78ffc940c4bb736df7336c5b3bf1
Author: Harmen Stoppels <harmenstoppels@gmail.com>
Date:   Sun Mar 20 22:41:54 2022 +0100

    environment.py: more defensive check before rm -rf

    Currently `old_root` is computed by reading the symlink at `self.root`.
    We should be more defensive in removing it by checking that it is in the
    same directory as the new root. Otherwise, in the worst case, when
    someone runs `spack env create --with-view=./view -d .` and `view`
    already exists and is a symlink to `/`, Spack effectively runs `rm -rf /`.

+++ lib/spack/spack/environment/environment.py
@@ -554,8 +554,14 @@ class ViewDescriptor(object):
                 pass
             raise e
 
-        # remove old_root
-        if old_root and os.path.exists(old_root):
+        # Remove the old root when it's in the same folder as the new root. This guards
+        # against removal of an arbitrary path when the original symlink in self.root
+        # was not created by the environment, but by the user.
+        if (
+            old_root and
+            os.path.exists(old_root) and
+            os.path.samefile(os.path.dirname(new_root), os.path.dirname(old_root))
+        ):
             try:
                 shutil.rmtree(old_root)
             except (IOError, OSError) as e:
